Added Delete Endpoint and Debug endpoint to populate database randomly.

+++ InventariumAPI/Repositories/ObjectRepository.cs
@@ -10,11 +10,20 @@ public class ObjectRepository(DataContext _context)
 {
     private readonly DataContext context = _context;
 
+    public Task AddManager(TModelId id, TModelId userId)
+        => Task.Run(() => context.ObjectManagers.AddAsync(new(){
+            ObjectId = id,
+            UserId = userId
+        }));
+
     public ValueTask<BrokenObject?> AsBrokenAsync(TModelId id)
         => context.BrokenObjects.FindAsync(id);
 
     public async Task<Lendout?> GetLendout(TModelId id)
         => await context.Lendouts
+            .IgnoreAutoIncludes()
+            .Include(l => l.User)
+            .Include(l => l.Object)
             .Where(l => l.ObjectId == id)
             .Where(l => l.StartDate <= DateTime.UtcNow 
                 && (l.EndDate == null 
@@ -25,33 +34,45 @@ public class ObjectRepository(DataContext _context)
         
     public async Task<IEnumerable<User>?> GetManagers(TModelId id)
         => await context.ObjectManagers
+            .IgnoreAutoIncludes()
+            .Include(l => l.User)
+            .Include(l => l.Object)
             .Where(m => m.ObjectId == id)
             .ToListAsync()
             .ContinueWith(l => l.Result.Select(m => m.User));
 
+    public async Task RemoveManager(int id, int userId)
+        => await context.ObjectManagers
+            .Where (m => m.ObjectId == id && m.UserId == userId)
+            .ExecuteDeleteAsync();
+
     public async Task<bool?> SetBrokenAsync(int id, string? reason)
     {
+
         if (!await DoesExistAsync(id))
             return null;
 
-        if (reason is not null)
+        var broken = await context.BrokenObjects.FindAsync(id);
+
+        if (broken is null && reason is null)
+            return null;
+        else if (reason is null && broken is not null)
         {
-            var broken = await context.BrokenObjects.FindAsync(id);
-
-            if (broken == null)
-                return null;
-
             context.BrokenObjects.Remove(broken);
-
-            return false;
+            return true;
+        } else if (reason is not null && broken is not null)
+        {
+            broken.Reason = reason!;
+            context.Update(broken);
+        } else if (reason is not null && broken is null)
+        {
+            await context.BrokenObjects.AddAsync(new BrokenObject()
+            {
+                ObjectId = id,
+                Reason = reason!
+            });
         }
 
-        context.BrokenObjects.Update(new BrokenObject()
-        {
-            ObjectId = id,
-            Reason = reason!
-        });
-
-        return true;
+        return false;
     }
 }
